Create JourneyTimeOptionsGroup
